Mise en place de GameInvite et ajout de fonctionnalité dans Admin et General Ajout de commentaire dans EventHandler
